feat: made db a plugin
