+++ backend/src/routes.ts
@@ -1,4 +1,3 @@
-import db from "./db";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
 
 /*
@@ -19,7 +18,7 @@ async function routes (fastify: FastifyInstance, options: FastifyPluginOptions) 
 	fastify.post("/users", async (request, reply) => {
 		const { username } = request.body as { username: string };
 		try {
-		  const stmt = db.prepare("INSERT INTO users (username) VALUES (?)");
+		  const stmt = fastify.db.prepare("INSERT INTO users (username) VALUES (?)");
 		  const info = stmt.run(username);
 
 		  return { id: info.lastInsertRowid, username };
@@ -37,14 +36,14 @@ async function routes (fastify: FastifyInstance, options: FastifyPluginOptions) 
 
 	  // READ (all users)
 	  fastify.get("/users", async () => {
-		const rows = db.prepare("SELECT * FROM users").all();
+		const rows = fastify.db.prepare("SELECT * FROM users").all();
 		return rows;
 	  });
 
 	  // READ (single user)
 	  fastify.get("/users/:id", async (request) => {
 		const { id } = request.params as { id: string };
-		const row = db.prepare("SELECT * FROM users WHERE id = ?").get(id);
+		const row = fastify.db.prepare("SELECT * FROM users WHERE id = ?").get(id);
 		return row || { error: "User not found" };
 	  });
 
@@ -52,7 +51,7 @@ async function routes (fastify: FastifyInstance, options: FastifyPluginOptions) 
 	  fastify.put("/users/:id", async (request, reply) => {
 		const { id } = request.params as { id: string };
 		const { username } = request.body as { username: string };
-		const stmt = db.prepare("UPDATE users SET username = ? WHERE id = ?");
+		const stmt = fastify.db.prepare("UPDATE users SET username = ? WHERE id = ?");
 		const info = stmt.run(username, id);
 		return info.changes > 0 ? { id, username } : { error: "User not found" };
 	  });
@@ -60,7 +59,7 @@ async function routes (fastify: FastifyInstance, options: FastifyPluginOptions) 
 	  // DELETE
 	  fastify.delete("/users/:id", async (request) => {
 		const { id } = request.params as { id: string };
-		const stmt = db.prepare("DELETE FROM users WHERE id = ?");
+		const stmt = fastify.db.prepare("DELETE FROM users WHERE id = ?");
 		const info = stmt.run(id);
 		return info.changes > 0 ? { id } : { error: "User not found" };
 	  });
