+++ backend/src/server.ts
@@ -1,25 +1,12 @@
-import Fastify from "fastify";
-import fastifyCors from "@fastify/cors";
-import routes from "./routes"
-
-const fastify = Fastify({
-  logger: true
-});
-
-fastify.register(fastifyCors, {
-	origin: "*",
-	methods: ["GET", "POST", "PUT", "DELETE"], // allow your methods
-  });
-
-fastify.register(routes);
+import app from "./app"
 
 // Start server
 const start = async () => {
   try {
-    await fastify.listen({ port: 3000, host: "0.0.0.0" });
+    await app.listen({ port: 3000, host: "0.0.0.0" });
     console.log("🚀 Server runnning at http://localhost:3000");
   } catch (err) {
-    fastify.log.error(err);
+    app.log.error(err);
     process.exit(1);
   }
 }
