+++ backend/src/types/fastify.d.ts
@@ -1,0 +1,10 @@
+// types/fastify.d.ts
+import "fastify";
+import Database from "better-sqlite3";
+
+// declare module is used to augment or extend the type definitions of an existing module (like fastify).
+declare module "fastify" {
+  interface FastifyInstance {
+    db: Database;
+  }
+}
